Extracts NNAPIDelegateKernel from nnapi_delegate.cc
PiperOrigin-RevId: 262571387
